refactor species service

+++ src/Imi.Project.Api.Core/Services/SpeciesService.cs
@@ -1,10 +1,11 @@
 ﻿using Imi.Project.Api.Core.Entities;
+using Imi.Project.Api.Core.Exceptions;
 using Imi.Project.Api.Core.Helper;
 using Imi.Project.Api.Core.Interfaces.Repositories;
 using Imi.Project.Api.Core.Interfaces.Services;
+using Imi.Project.Common.Dtos;
 using System;
 using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace Imi.Project.Api.Core.Services
@@ -18,22 +19,33 @@ namespace Imi.Project.Api.Core.Services
             _speciesRepository = speciesRepository;
         }
 
-        public async Task<Species> AddSpeciesAsync(Species species)
+        public async Task<SpeciesResponseDto> GetSpeciesByIdAsync(Guid id)
         {
-
-            var result = await _speciesRepository.AddAsync(species);
+            var species = await _speciesRepository.GetByIdAsync(id);
+            if (species == null)
+            {
+                throw new BadRequestException($"Species with id {id} does not exist");
+            }
+            SpeciesResponseDto result = species.MapToDto();
             return result;
         }
 
-        public async Task DeleteSpeciesAsync(Species species)
+        public async Task<SpeciesResponseDto> AddSpeciesAsync(SpeciesRequestDto dto)
         {
-            await _speciesRepository.DeleteAsync(species);       
+            var newSpeciesEntity = dto.MapToEntity();
+            var result = await _speciesRepository.AddAsync(newSpeciesEntity);
+            var resultDto = result.MapToDto();
+            return resultDto;
         }
 
-        public async Task<Species> GetSpeciesByIdAsync(Guid id)
+        public async Task DeleteSpeciesAsync(Guid id)
         {
             var species = await _speciesRepository.GetByIdAsync(id);
-            return species;
+            if (species == null)
+            {
+                throw new BadRequestException($"Species with id {id} does not exist");
+            }
+            await _speciesRepository.DeleteAsync(species);
         }
 
         public async Task<IEnumerable<Species>> ListAllSpeciessAsync()
@@ -42,11 +54,17 @@ namespace Imi.Project.Api.Core.Services
             return species;
         }
 
-        public async Task<Species> UpdateSpeciesAsync(Species species)
+        public async Task<SpeciesResponseDto> UpdateSpeciesAsync(Guid id, SpeciesRequestDto dto)
         {
-            var result = await _speciesRepository.UpdateAsync(species);
-            return result;
-
+            var species = await _speciesRepository.GetByIdAsync(id);
+            if (species == null)
+            {
+                throw new BadRequestException($"Species with id {id} does not exist");
+            }
+            var newSpeciesEntity = dto.MapToEntity();
+            var result = await _speciesRepository.AddAsync(newSpeciesEntity);
+            var resultDto = result.MapToDto();
+            return resultDto;
         }
     }
 }
